redis -- abandoned
This is pretty darn fast -- prelim tests indicate 100x-200x speed up on
the access() function call using redis.  This is on Fedora's humungous
conf file.  (Redis took about 30 MB of RAM, which is not a big deal).

There's a good bit of work left in order to completely convert it.  It's
not that hard, and should actually be fun so I *may* do it sometime.

But it's not really worth it, since the ssh+network overhead completely
overshadows the time taken for the access() call anyway.  (And don't
forget git itself!)